localization_translator/requirements.txt
PyQt5
google-cloud-translate
openai
pandas
